token_authed_fetch and cacheAll share Request format.
